Create only valid QDateTime's.
With an invalid QTime aka QTime() the QDateTime is also invalid. This
hack was used to have something like a QDateTime with a dateonly feature.

With this patch we no longer create invalid date-times, at the expense of
converting all date-only brithdays to the time 00:00:00. If support for
date-only birthdays is indeed required, the interface needs to be adjusted
or QDateTime needs to gain a date-only feature.

KOLAB: 3935

Conflicts:
	kabc/vcardtool.cpp