Addressee: Determine if the birthday has been set with a time
Implemented a way to determine if there has been set a explicit time
with the date of birth and that the time is only exported to a vcard
if it has been explicitly specified. Some vcard implementations (e.g.
on Android) ignore the birthday field if it comes with a time.

The problem here is that QDateTime (when set with a valid date) always
has a valid time (midnight by default). I therefore added the boolean
withTime to setBirthday() which is stored inside Addressee and can be
queried with birthdayHasTime(). I also added setBirthday(QDate).

I changed VCardTool and LDIFConverter to make use of this new methods
when importing/exporting contacts.

Unrelated to this changes I implemented a way in VCardTool to import/
export dates without a year. Dates without a year ('--MMdd') are
imported as year=-1 and during export, years <= 0 are written as '--'.
Also added import/export without year to LDIFConverter.

REVIEW: 128743