Fix translation of country names
Summary:
There is no catalog containing the country names right now, so calling
i18n() with country names has no effect. The iso-codes package provides
catalogs for this, so let's use that.

One could argue that this violates the dependency freeze, however it will
not change anything if the iso-codes package isn't present, you can still
build and get untranslated country names. It does however fix the issue
when iso-codes happens to be present, which is likely considering Plasma
depends on it already.

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D14163