Add CLDR-based country to ISO code mapping
Summary:
This replaces the many year outdated countrytransl.map file with a
compiled-in data table. This is much more complete (28k strings rather
than the previous ~5k), up-to-date (the old data set was from a time
when Yugoslavia was still a thing), and much much faster to access
(no more text file parsing for each lookup, but a binary search in
shared read-only data). However, we pay for this by needing more disk
space (670kB vs 95kB), which is largely due to the increased amount of
strings, many of which are in "expensive" encodings, the overhead per
string only increases from four to five bytes.

This is based on Sune's work in kde:scratch/sune/countrieslistcldr.

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D14563