Remove the configurable operator< behavior
Summary:
Besides being unused, this is very dangerous as changes to the sort mode
would affect every single user of this without a chance to prevent this.

This also removes the entire opertor< for Addressee, which for such a
complex class actually makes sense, as there are many many different
ways to sort Addressees.

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D20273