Improve handling of clashing country name strings
Summary:
This now no longer removes name clashes within one language by using a
hash (and yes, those cases do seem to exist), and it propagates the
language information all the way to the error message to make it easier
to investigate those issues.

This is mainly a preparation for applying more aggressive string
normalization on the country names, which we can do as along is it
doesn't increase the amount of ambiguities here.

Reviewers: dvratil

Reviewed By: dvratil

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D20415