Make IMPP more compliant with RFC 4770/ RFC 6350
Summary:
The payload of IMPP entries is supposed to be a URI, the scheme of which
identifies the corresponding messaging service. This can still read the
pre-RFC 4770 service type parameter if no URI scheme is given, but it now
always produces URIs in the output. With RFC 4770 being twelve years old
already I think that's acceptable.

The removal of the type() method also opens the way for adding API for
the type parameter as we have it for other fields as well.

Note that this does not yet touch pre-IMPP messaging entries as currently
used by KAddressbook, that still has to be changed. This also means that
we currently have no users of this code, making these rather invasive
changes possible.

Reviewers: mlaurent

Reviewed By: mlaurent

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D20706