Compile benchmark tests, but don't run them automatically with ctest
Benchmark unit tests are not useful if no one is looking at the results
(either when run locally or on the CI); change it to be always compiled,
so as not to bit-rot, but don't run them automatically when using e.g.
`ctest` or `make test`.

See https://invent.kde.org/frameworks/kio/-/merge_requests/301