Addressee: deprecate insertEmail() and add a new method instead
addEmail() takes an Email object, i.e. you create an Email, call the setter
methods to make it preferred, set the type ...etc, then add it to Addressee.

Fix unit test: the code in insertEmail() only considered the bool "preferred"
function parameter, but not the perferred parameter in the parameter map
passed to insertEmail(). Reading the code the "preferred" email should be
first in the list returned by Adressee::emails().