Deprecate parameters/setParameters() from the public API
I've had a long chat with Volker:
- we found the parameters setter/getter to be only used in libkgapi (which
  has since been ported to use the setType()/type() methods).
- the vcard format parsing is really an implementation detail
  that needn't be leaked to the public API
- Wrap the "parameter map" logic in one class, that has an API similar to
  a map, so ParameterMap is added.
Thanks to Volker for the API design (I mostly did the leg work :)).

The plan is to deprecate the public methods to keep backwards compatibility
in KF5, and add new private methods that do the same job, for the KContacts
code that needs that functionality, and then those classes can be made friends.

The QDataStream serializing is based on the implemention of QMap serialization
in QDataStream.