Use local country names when formatting addresses
As this is aimed at postal addresses and we use the local formatting
rule for the country part, also using the local country name would be
consistent.

Also, this now properly expands ISO 3166-1 codes in the country field for
address formatting, which is something a number of consumers are doing
manually right now.