Improve output for incomplete addresses and formats with 3+ fields per line
This deviates from the libaddressinput algorithm in that we now also
include literals if any preceding field in the same line is non-empty,
not just the immediately preceding one. This doesn't change anything
when all required fields are provided in the input address, but it
ensures we get proper output for e.g. "%C %S %Z" if region and postal
code are insufficiently separated in the address data.