Addressee: make birthday writable from QML
Add a private Addressee::setBirthdayProperty(const QDateTime &) wrapper
method that guesses the extra withTime argument and use that as the
WRITE method for the Q_PROPERTY birthday.

A non-midnight time spec will lead to the birthday being recorded with
the time, a time spec exactly at midnight to a date-only birthday (the
common case). This is the best we can do for a property setter.

This should make setting the birthday in Plasma Phonebook work again.
(It last worked when Plasma Phonebook had its own QML wrapper for the
Addressee class, which has been dropped since.)

See also:
https://invent.kde.org/plasma-mobile/plasma-phonebook/-/issues/6#note_307938

BUG: 446665